Rename scope to @internationalized
