+++ packages/@internationalized/message/src/MessageDictionary.ts
