+++ packages/@internationalized/number/README.md
@@ -1,3 +1,3 @@
-# @spectrum-i18n/number
+# @internationalized/number
 
 This package is part of [react-spectrum](https://github.com/adobe/react-spectrum). See the repo for more details.
