+++ packages/@internationalized/number/src/NumberParser.ts
