+++ packages/@react-aria/i18n/package.json
@@ -21,8 +21,8 @@
     "@react-aria/ssr": "^3.0.1",
     "@react-aria/utils": "^3.4.0",
     "@react-types/shared": "^3.3.0",
-    "@spectrum-i18n/message": "3.0.0-alpha.1",
-    "@spectrum-i18n/number": "3.0.0-alpha.1"
+    "@internationalized/message": "3.0.0-alpha.1",
+    "@internationalized/number": "3.0.0-alpha.1"
   },
   "peerDependencies": {
     "react": "^16.8.0 || ^17.0.0-rc.1"
