+++ packages/@react-stately/numberfield/package.json
@@ -22,7 +22,7 @@
     "@react-stately/utils": "^3.0.2",
     "@react-types/numberfield": "3.0.0-alpha.1",
     "@react-types/shared": "^3.1.0",
-    "@spectrum-i18n/number": "3.0.0-alpha.1"
+    "@internationalized/number": "3.0.0-alpha.1"
   },
   "peerDependencies": {
     "react": "^16.8.0 || ^17.0.0-rc.1"
